+++ assets/javascripts/kitten/components/cards/team-card/test.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import renderer from 'react-test-renderer'
-import { TeamCard } from './team-card'
+import { TeamCard } from './index'
 
 const createMockMediaMatcher = matches => () => ({
   matches,